git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@79 dfa36f5b-b2fa-0310-9605-e5aca7e64462

+++ Source/wali/Key.hpp
@@ -20,7 +20,7 @@ namespace wali
      *
      * For now a Key is just an unsigned int
      */
-    typedef unsigned int Key;
+    typedef size_t Key;
 
     /*!
      * Defined in KeySource.hpp
